requirements.txt
docker==6.1.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
python-telegram-bot==20.3
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
websocket-client==1.5.2